+++ Applications/AdminApi/src/AdminApi/appsettings.json
@@ -1,91 +1,94 @@
 {
-  "AllowedHosts": "*",
-  "Authentication": {
-    "JwtLifetimeInSeconds": 300
-  },
-  "Infrastructure": {
-    "EventBus": {
-      "SubscriptionClientName": "adminui"
+    "AllowedHosts": "*",
+    "Authentication": {
+        "JwtLifetimeInSeconds": 300
+    },
+    "Infrastructure": {
+        "EventBus": {
+            "SubscriptionClientName": "adminui"
+        }
+    },
+    "Modules": {
+        "Announcements": {
+            "Application": {
+                "Pagination": {
+                    "DefaultPageSize": 50,
+                    "MaxPageSize": 200
+                }
+            },
+            "Infrastructure": {
+                "SqlDatabase": {
+                    "EnableHealthCheck": true
+                }
+            }
+        },
+        "Challenges": {
+            "Infrastructure": {
+                "SqlDatabase": {
+                    "EnableHealthCheck": true
+                }
+            }
+        },
+        "Quotas": {
+            "Application": {
+                "Pagination": {
+                    "DefaultPageSize": 50,
+                    "MaxPageSize": 200
+                }
+            },
+            "Infrastructure": {
+                "SqlDatabase": {
+                    "EnableHealthCheck": true
+                }
+            }
+        },
+        "Devices": {
+            "Application": {
+                "Pagination": {
+                    "DefaultPageSize": 50,
+                    "MaxPageSize": 200
+                }
+            },
+            "Infrastructure": {
+                "SqlDatabase": {
+                    "EnableHealthCheck": true
+                }
+            }
+        },
+
+        "Tokens": {
+            "Application": {
+                "Pagination": {
+                    "DefaultPageSize": 50,
+                    "MaxPageSize": 200
+                }
+            },
+            "Infrastructure": {
+                "SqlDatabase": {
+                    "EnableHealthCheck": true
+                }
+            }
+        }
+    },
+    "Logging": {
+        "MinimumLevel": {
+            "Default": "Warning",
+            "Override": {
+                "Backbone": "Information",
+                "Enmeshed": "Information",
+                "AdminApi": "Information",
+                "Microsoft": "Information"
+            }
+        },
+        "WriteTo": {
+            "Console": {
+                "Name": "Console",
+                "Args": {
+                    "formatter": "Serilog.Formatting.Compact.CompactJsonFormatter, Serilog.Formatting.Compact"
+                }
+            }
+        }
+
     }
-  },
-  "Modules": {
-    "Announcements": {
-      "Application": {
-        "Pagination": {
-          "DefaultPageSize": 50,
-          "MaxPageSize": 200
-        }
-      },
-      "Infrastructure": {
-        "SqlDatabase": {
-          "EnableHealthCheck": true
-        }
-      }
-    },
-    "Challenges": {
-      "Infrastructure": {
-        "SqlDatabase": {
-          "EnableHealthCheck": true
-        }
-      }
-    },
-    "Quotas": {
-      "Application": {
-        "Pagination": {
-          "DefaultPageSize": 50,
-          "MaxPageSize": 200
-        }
-      },
-      "Infrastructure": {
-        "SqlDatabase": {
-          "EnableHealthCheck": true
-        }
-      }
-    },
-    "Devices": {
-      "Application": {
-        "Pagination": {
-          "DefaultPageSize": 50,
-          "MaxPageSize": 200
-        }
-      },
-      "Infrastructure": {
-        "SqlDatabase": {
-          "EnableHealthCheck": true
-        }
-      }
-    },
-    "Tokens": {
-      "Application": {
-        "Pagination": {
-          "DefaultPageSize": 50,
-          "MaxPageSize": 200
-        }
-      },
-      "Infrastructure": {
-        "SqlDatabase": {
-          "EnableHealthCheck": true
-        }
-      }
-    }
-  },
-  "Logging": {
-    "MinimumLevel": {
-      "Default": "Warning",
-      "Override": {
-        "Backbone": "Information",
-        "Enmeshed": "Information",
-        "AdminApi": "Information",
-        "Microsoft": "Information"
-      }
-    },
-    "WriteTo": {
-      "Console": {
-        "Name": "Console",
-        "Args": {
-          "formatter": "Serilog.Formatting.Compact.CompactJsonFormatter, Serilog.Formatting.Compact"
-        }
-      }
-    }
-  }
+
 }
